Make @QuarkusMainTest respect `quarkus.test.profile.tags`
Fixes: #44118
(cherry picked from commit a82eba2b70152f717cab73f8fafecaee18cde7cc)
